refactor: remove old TCP server, move DbCache to http_server
The TCP wire protocol server (server.rs) is fully superseded by the
HTTP/2 server. Move DbCache into http_server.rs and delete server.rs.
