+++ src/lib.rs
@@ -30,6 +30,5 @@ pub mod config;
 pub mod http_server;
 pub mod node;
 pub mod protocol;
-pub mod server;
 
 pub use node::{Database, IntoQuery, Node, QueryNode, SubmitNode, TransactionResult, TxKey, DB};
